+++ pkg/teler/teler.go
@@ -11,13 +11,13 @@ import (
 	"github.com/valyala/fastjson"
 	"ktbs.dev/teler/common"
 	"ktbs.dev/teler/pkg/matchers"
+	"ktbs.dev/teler/pkg/metrics"
 	"ktbs.dev/teler/resource"
 )
 
 // Analyze logs from threat resources
 func Analyze(options *common.Options, logs *gonx.Entry) (bool, map[string]string) {
 	var match bool
-
 	log := make(map[string]string)
 	rsc := resource.Get()
 
@@ -33,6 +33,7 @@ func Analyze(options *common.Options, logs *gonx.Entry) (bool, map[string]string
 		exc := threat.FieldByName("Exclude").Bool()
 
 		log["category"] = cat
+		metrics.GetThreasTotal.WithLabelValues(cat).Inc()
 
 		if exc {
 			continue
@@ -65,30 +66,44 @@ func Analyze(options *common.Options, logs *gonx.Entry) (bool, map[string]string
 							quote,
 						)
 						if match {
+							metrics.GetCWA.WithLabelValues(string(v.GetStringBytes("description")),
+								log["remote_addr"], log["request_uri"], log["status"]).Inc()
+
 							break
 						}
 					}
 				}
 			}
+
 		case "Bad Crawler":
 			log["element"] = "http_user_agent"
+
 			if white := isWhitelist(options, log["http_user_agent"]); white {
 				break
 			}
 
 			for _, pat := range strings.Split(con, "\n") {
 				if match = matchers.IsMatch(pat, log["http_user_agent"]); match {
+
+					metrics.GetBadCrawler.WithLabelValues(log["remote_addr"],
+						log["http_user_agent"],
+						log["status"]).Inc()
+
 					break
 				}
 			}
+
 		case "Bad IP Address":
 			log["element"] = "remote_addr"
+
 			if white := isWhitelist(options, log["remote_addr"]); white {
 				break
 			}
 
 			ip := "(?m)^" + log["remote_addr"]
 			match = matchers.IsMatch(ip, con)
+			metrics.GetBadIP.WithLabelValues(log["remote_addr"]).Inc()
+
 		case "Bad Referrer":
 			log["element"] = "http_referer"
 			if white := isWhitelist(options, log["http_referer"]); white {
@@ -102,8 +117,11 @@ func Analyze(options *common.Options, logs *gonx.Entry) (bool, map[string]string
 			ref := "(?m)^" + req.Host
 
 			match = matchers.IsMatch(ref, con)
+			metrics.GetBadReferrer.WithLabelValues(log["http_referer"]).Inc()
+
 		case "Directory Bruteforce":
 			log["element"] = "request_uri"
+
 			if white := isWhitelist(options, log["request_uri"]); white {
 				break
 			}
@@ -122,6 +140,9 @@ func Analyze(options *common.Options, logs *gonx.Entry) (bool, map[string]string
 				case "200", "204", "304":
 					match = false
 				}
+				metrics.GetDirBruteforce.WithLabelValues(log["remote_addr"],
+					log["request_uri"],
+					log["status"]).Inc()
 			}
 		}